Разделил код проекта на модули, настроил сборку проекта через вебпак, код минифицируется и адаптируется для старых браузеров.
