+++ postcss.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    plugins: [
+        require('autoprefixer'),
+        require('cssnano')({ // подключили cssnano
+            preset: 'default', // выбрали настройки по умолчанию
+    })
+]
+};
